Rename handlers directory to be more generic.
